+++ R/read-data.R
@@ -3,7 +3,7 @@
 #' It reads the raw data exported from the metabolic cart.
 #'
 #' @param path Path to read the file from.
-#' @param metabolic_cart Metabolic cart that was used for data collection. Currently, 'cosmed', 'cortex', 'nspire', 'parvo', and 'geratherm' are supported.
+#' @param metabolic_cart Metabolic cart that was used for data collection. Currently, 'cosmed', 'cortex', 'nspire', 'parvo', 'geratherm', and 'cardiocoach' are supported.
 #' @param time_column The name (quoted) of the column containing the time. Depending on the language of your system, this column might not be "t". Therefore, you may specify it here.  Default to "t".
 #' @param work_rate_column Default is `NULL`. In case your work rate column is coerced as a character column
 #' you can define here the name of this column in your data file. This happens because at the very beginning of the test
@@ -15,7 +15,7 @@
 #' @export
 read_data <- function(
   path,
-  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm"),
+  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm", "cardiocoach"),
   time_column = "t",
   work_rate_column = NULL
 ) {
@@ -32,7 +32,7 @@ read_data <- function(
 #' @export
 read_data.cosmed <- function(
   path,
-  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm"),
+  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm", "cardiocoach"),
   time_column = "t",
   work_rate_column = NULL
 ) {
@@ -112,7 +112,7 @@ read_data.cosmed <- function(
 #' @export
 read_data.cortex <- function(
   path,
-  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm"),
+  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm", "cardiocoach"),
   time_column = "t",
   work_rate_column = NULL
 ) {
@@ -178,7 +178,7 @@ read_data.cortex <- function(
 #' @export
 read_data.nspire <- function(
   path,
-  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm"),
+  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm", "cardiocoach"),
   time_column = "t",
   work_rate_column = NULL
 ) {
@@ -226,7 +226,7 @@ read_data.nspire <- function(
 #' @export
 read_data.parvo <- function(
   path,
-  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm"),
+  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm", "cardiocoach"),
   time_column = "t",
   work_rate_column = NULL
 ) {
@@ -280,7 +280,7 @@ read_data.parvo <- function(
 #' @export
 read_data.geratherm <- function(
   path,
-  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm"),
+  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm", "cardiocoach"),
   time_column = "t",
   work_rate_column = NULL
 ) {
@@ -332,6 +332,55 @@ read_data.geratherm <- function(
   metadata$metabolic_cart <- "Geratherm"
   metadata$data_status <- "raw data"
   metadata$time_column <- time_column
+
+  out <- new_whippr_tibble(out, metadata)
+
+  out
+}
+
+#' @export
+read_data.cardiocoach <- function(
+  path,
+  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo", "geratherm", "cardiocoach"),
+  time_column = "t",
+  work_rate_column = NULL
+) {
+
+  ## retrieve column names
+  names_file <- suppressMessages(readr::read_tsv(file = path, n_max = 2, col_names = FALSE))
+  ## retrieve data
+  readings <- suppressMessages(readr::read_tsv(file = path, col_names = FALSE, skip = 2))
+
+  column_names <- dplyr::tibble(
+    X1 = unlist(c(names_file[1,])),
+    X2 = unlist(c(names_file[2,]))
+  ) %>%
+    dplyr::mutate(
+      names = ifelse(is.na(X2), X1, glue::glue("{X1} ({X2})"))
+    ) %>%
+    dplyr::pull(names)
+
+  ## define time column as 't' for simplicity
+  column_names[1] <- "t"
+
+  colnames(readings) <- column_names
+
+  out <- readings %>%
+    dplyr::mutate(
+      t = dplyr::case_when(
+        nchar(t) == 4 ~ paste0("00:0", t),
+        nchar(t) == 5 ~ paste0("00:", t),
+        TRUE ~ t
+      ),
+      t = lubridate::hms(t),
+      t = lubridate::period_to_seconds(t)
+    )
+
+  metadata <- NULL
+  metadata$read_data <- TRUE
+  metadata$metabolic_cart <- "CardioCoach"
+  metadata$data_status <- "raw data"
+  metadata$time_column <- "t"
 
   out <- new_whippr_tibble(out, metadata)